+++ src/ui/mod.rs
@@ -1,3 +1,3 @@
 pub mod args;
 pub mod cli;
-pub mod id;
+pub mod idshortcuts;